agent/requirements.txt
livekit-agents>=1.0
livekit-plugins-deepgram>=1.0
livekit-plugins-cartesia>=1.0
livekit-plugins-anthropic>=1.0
supabase>=2.0
python-dotenv>=1.0
